emu/video: Got rid of all RGB utilities specialisations besides bilinear filtering, added ARM NEON bilinear filtering.
