+++ scripts/src/emu.lua
@@ -267,13 +267,8 @@ files {
 	MAME_DIR .. "src/emu/video/generic.h",
 	MAME_DIR .. "src/emu/video/resnet.cpp",
 	MAME_DIR .. "src/emu/video/resnet.h",
+	MAME_DIR .. "src/emu/video/rgbutil.cpp",
 	MAME_DIR .. "src/emu/video/rgbutil.h",
-	MAME_DIR .. "src/emu/video/rgbgen.cpp",
-	MAME_DIR .. "src/emu/video/rgbgen.h",
-	MAME_DIR .. "src/emu/video/rgbsse.cpp",
-	MAME_DIR .. "src/emu/video/rgbsse.h",
-	MAME_DIR .. "src/emu/video/rgbvmx.cpp",
-	MAME_DIR .. "src/emu/video/rgbvmx.h",
 	MAME_DIR .. "src/emu/audio_effects/aeffect.h",
 	MAME_DIR .. "src/emu/audio_effects/aeffect.cpp",
 	MAME_DIR .. "src/emu/audio_effects/filter.h",
